+++ app/src/main/java/com/codesoom/assignment/application/BasicTaskService.java
@@ -2,7 +2,7 @@ package com.codesoom.assignment.application;
 
 import com.codesoom.assignment.TaskNotFoundException;
 import com.codesoom.assignment.models.Task;
-import com.codesoom.assignment.repository.TaskRepository;
+import com.codesoom.assignment.repository.TaskMapRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -10,23 +10,23 @@ import java.util.Collection;
 
 @Service
 public class BasicTaskService implements TaskService{
-    private final TaskRepository taskRepository;
+    private final TaskMapRepository taskMapRepository;
 
-    public BasicTaskService(TaskRepository taskRepository) {
-        this.taskRepository = taskRepository;
+    public BasicTaskService(TaskMapRepository taskMapRepository) {
+        this.taskMapRepository = taskMapRepository;
     }
 
     public Collection<Task> getTasks() {
-        return new ArrayList<>(taskRepository.getAll());
+        return new ArrayList<>(taskMapRepository.getAll());
     }
 
     public Task getTask(Long id) {
-        return taskRepository.get(id)
+        return taskMapRepository.get(id)
                     .orElseThrow(() -> new TaskNotFoundException(id));
     }
 
     public Task createTask(String title) {
-        return taskRepository.add(title);
+        return taskMapRepository.add(title);
     }
 
     public Task updateTask(Long id, String title) {
@@ -35,6 +35,6 @@ public class BasicTaskService implements TaskService{
 
     public void deleteTask(Long id) {
         getTask(id);
-        taskRepository.remove(id);
+        taskMapRepository.remove(id);
     }
 }
